fix: save event on publish
