+++ src/app/main/component/events/components/event-details/event-details.component.html
@@ -30,11 +30,9 @@
           </button>
         </div>
         <div *ngIf="isPreview">
-          <!-- <a [routerLink]="['/events', 'create-event']" class="button-link"> -->
           <button class="primary-global-button button-publish" (click)="onPublish()">
             {{ 'create-event.publish' | translate }}
           </button>
-          <!-- </a> -->
         </div>
       </div>
