+++ src/app/main/component/events/components/event-editor/event-editor.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
-import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBarComponent } from '@global-errors/mat-snack-bar/mat-snack-bar.component';
@@ -17,7 +17,7 @@ import { CreateEcoEventAction, EditEcoEventAction, EventsActions } from 'src/app
 import { singleNewsImages } from 'src/app/main/image-pathes/single-news-images';
 import { Place } from '../../../places/models/place';
 import { DefaultCoordinates } from '../../models/event-consts';
-import { DateInformation, Dates, EventDTO, EventForm, EventResponse, LocationResponse, TagObj } from '../../models/events.interface';
+import { EventForm, EventResponse, TagObj } from '../../models/events.interface';
 import { EventsService } from '../../services/events.service';
 import { quillConfig } from './quillEditorFunc';
 import { EventStoreService } from '../../services/event-store.service';
@@ -34,33 +34,13 @@ export class EventEditorComponent extends FormBaseComponent implements OnInit {
   quillModules = {};
   places: Place[] = [];
   isPosting = false;
-  editEvent: EventResponse;
-  tags: Array<TagObj>;
-  isImageSizeError: boolean;
-  isImageTypeError = false;
   images = singleNewsImages;
-  currentLang: string;
   submitButtonName = 'create-event.publish';
   subscription: Subscription;
-  imgArray: Array<File> = [];
-  userId: number;
   previousPath: string;
   eventForm: FormGroup;
   event: EventForm;
-  popupConfig = {
-    hasBackdrop: true,
-    closeOnNavigation: true,
-    disableClose: true,
-    panelClass: 'popup-dialog-container',
-    data: {
-      popupTitle: 'homepage.events.events-popup.title',
-      popupSubtitle: 'homepage.events.events-popup.subtitle',
-      popupConfirm: 'homepage.events.events-popup.confirm',
-      popupCancel: 'homepage.events.events-popup.cancel'
-    }
-  };
   routedFromProfile: boolean;
-  private _savedFormValues: EventForm;
 
   constructor(
     public dialog: MatDialog,
@@ -102,7 +82,7 @@ export class EventEditorComponent extends FormBaseComponent implements OnInit {
     if (this.isUpdating) {
       this.submitButtonName = 'create-event.save-event';
     }
-    this.createFormEvent();
+    this.eventForm = this.eventsService.convertEventToFormEvent(this.event);
     this.routedFromProfile = this.localStorageService.getPreviousPage() === '/profile';
     this.previousPath = this.localStorageService.getPreviousPage() || '/events';
     this.eventForm
@@ -173,46 +153,6 @@ export class EventEditorComponent extends FormBaseComponent implements OnInit {
     });
   }
 
-  private createFormEvent(): void {
-    const information = this.event?.eventInformation;
-    const date = this.event?.dateInformation ?? [];
-
-    this.eventForm = this.fb.group({
-      eventInformation: this.fb.group({
-        title: [information?.title ?? '', [Validators.required, Validators.maxLength(70)]],
-        description: [information?.description ?? '', [Validators.required, Validators.minLength(20)]],
-        open: [information?.open ?? true, Validators.required],
-        images: [information?.images ?? []],
-        duration: [information?.duration ?? 1, Validators.required],
-        tags: [information?.tags ?? [], [Validators.required, Validators.minLength(1)]]
-      }),
-
-      dateInformation: this.fb.array(date.length > 0 ? date.map((date) => this.createDateFormGroup(date)) : [this.createDateFormGroup()])
-    });
-  }
-
-  private createDateFormGroup(date?: any): FormGroup {
-    return this.fb.group({
-      day: this.fb.group({
-        date: [date?.day.date ? new Date(date.day.date) : new Date(), [Validators.required]],
-        startTime: [date?.day.startTime ?? '', Validators.required],
-        endTime: [date?.day.endTime ?? '', Validators.required],
-        allDay: [date?.day.allDay ?? false],
-        minDate: [date?.day.minDate ? new Date(date.minDate) : new Date()],
-        maxDate: [date?.day.maxDate ? new Date(date.maxDate) : '']
-      }),
-      placeOnline: this.fb.group({
-        coordinates: new FormControl(
-          date?.placeOnline.coordinates ?? { lat: DefaultCoordinates.LATITUDE, lng: DefaultCoordinates.LONGITUDE }
-        ),
-        onlineLink: new FormControl(date?.placeOnline.onlineLink ?? ''),
-        place: new FormControl(date?.placeOnline.place ?? ''),
-        appliedLinkForAll: [date?.placeOnline.appliedLinkForAll ?? false],
-        appliedPlaceForAll: [date?.placeOnline.appliedPlaceForAll ?? false]
-      })
-    });
-  }
-
   onPreview() {
     const currentRoute = this.router.url;
     this.cdRef.detectChanges();
@@ -228,90 +168,12 @@ export class EventEditorComponent extends FormBaseComponent implements OnInit {
   }
 
   submitEvent(): void {
-    const { eventInformation, dateInformation } = this.eventForm.value;
-    const { open, tags, description, title, images } = eventInformation;
-    const dates: Dates[] = this.transformDatesFormToDates(dateInformation);
-    let sendEventDto: EventDTO = {
-      title,
-      description: description,
-      open,
-      tags,
-      datesLocations: dates
-    };
-
-    if (this.isUpdating) {
-      if (!this.eventId) {
-        const urlSegments = this.router.url.split('/');
-        this.eventId = Number(urlSegments[urlSegments.length - 1]);
-      }
-      const currentImages = (this._savedFormValues?.eventInformation?.images || [])
-        .filter((value) => !value.file)
-        .map((value) => value.url);
-      sendEventDto = {
-        ...sendEventDto,
-        additionalImages: currentImages.slice(1),
-        id: this.eventId,
-        titleImage: currentImages[0]
-      };
-    }
-    const formData: FormData = new FormData();
-    const stringifyDataToSend = JSON.stringify(sendEventDto);
-    const dtoName = this.isUpdating ? 'eventDto' : 'addEventDtoRequest';
-
-    formData.append(dtoName, stringifyDataToSend);
-
-    images.forEach((item) => {
-      if (item.file) {
-        formData.append('images', item.file);
-      }
-    });
-
+    const formData = this.eventsService.prepareEventForSubmit(this.eventForm.value, this.eventId, this.isUpdating);
     this.createEvent(formData);
   }
 
   clear(): void {
     this.eventForm.reset();
-  }
-
-  transformDatesFormToDates(form: DateInformation[]): Dates[] {
-    return form
-      .map((value) => {
-        const { date, endTime, startTime } = value.day;
-        const { onlineLink, place, coordinates } = value.placeOnline;
-
-        const dateObject = new Date(date);
-
-        if (isNaN(dateObject.getTime())) {
-          return;
-        }
-
-        let [hours, minutes] = startTime.split(':');
-        dateObject.setHours(parseInt(hours, 10));
-        dateObject.setMinutes(parseInt(minutes, 10));
-        const startDate = dateObject.toISOString();
-
-        [hours, minutes] = endTime.split(':');
-        dateObject.setHours(parseInt(hours, 10));
-        dateObject.setMinutes(parseInt(minutes, 10));
-        const finishDate = dateObject.toISOString();
-
-        const dates: Dates = {
-          startDate,
-          finishDate,
-          id: undefined
-        };
-        if (onlineLink) {
-          dates.onlineLink = onlineLink;
-        }
-        if (place) {
-          dates.coordinates = {
-            latitude: coordinates.lat,
-            longitude: coordinates.lng
-          };
-        }
-        return dates;
-      })
-      .filter(Boolean);
   }
 
   escapeFromCreateEvent(): void {
