+++ src/app/main/component/events/services/events.service.ts
@@ -2,9 +2,14 @@ import { Injectable, OnDestroy } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable, ReplaySubject } from 'rxjs';
 import { environment } from '@environment/environment';
+import { DefaultCoordinates } from '../models/event-consts';
+
 import {
   Addresses,
+  DateInformation,
+  Dates,
   EventAttender,
+  EventDTO,
   EventForm,
   EventResponse,
   EventResponseDto,
@@ -12,6 +17,7 @@ import {
   PagePreviewDTO
 } from '../models/events.interface';
 import { LanguageService } from 'src/app/main/i18n/language.service';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 
 @Injectable({
   providedIn: 'root'
@@ -26,7 +32,8 @@ export class EventsService implements OnDestroy {
 
   constructor(
     private http: HttpClient,
-    private langService: LanguageService
+    private langService: LanguageService,
+    private fb: FormBuilder
   ) {}
 
   setIsFromCreateEvent(value: boolean): void {
@@ -76,6 +83,123 @@ export class EventsService implements OnDestroy {
         .filter((place) => place)
         .join(', ')
     };
+  }
+
+  convertEventToFormEvent(event: EventForm): FormGroup {
+    const information = event?.eventInformation;
+    const date = event?.dateInformation ?? [];
+
+    return this.fb.group({
+      eventInformation: this.fb.group({
+        title: [information?.title ?? '', [Validators.required, Validators.maxLength(70)]],
+        description: [information?.description ?? '', [Validators.required, Validators.minLength(20)]],
+        open: [information?.open ?? true, Validators.required],
+        images: [information?.images ?? []],
+        duration: [information?.duration ?? 1, Validators.required],
+        tags: [information?.tags ?? [], [Validators.required, Validators.minLength(1)]]
+      }),
+
+      dateInformation: this.fb.array(date.length > 0 ? date.map((date) => this.createDateFormGroup(date)) : [this.createDateFormGroup()])
+    });
+  }
+
+  private createDateFormGroup(date?: any): FormGroup {
+    return this.fb.group({
+      day: this.fb.group({
+        date: [date?.day.date ? new Date(date.day.date) : new Date(), [Validators.required]],
+        startTime: [date?.day.startTime ?? '', Validators.required],
+        endTime: [date?.day.endTime ?? '', Validators.required],
+        allDay: [date?.day.allDay ?? false],
+        minDate: [date?.day.minDate ? new Date(date.minDate) : new Date()],
+        maxDate: [date?.day.maxDate ? new Date(date.maxDate) : '']
+      }),
+      placeOnline: this.fb.group({
+        coordinates: new FormControl(
+          date?.placeOnline.coordinates ?? { lat: DefaultCoordinates.LATITUDE, lng: DefaultCoordinates.LONGITUDE }
+        ),
+        onlineLink: new FormControl(date?.placeOnline.onlineLink ?? ''),
+        place: new FormControl(date?.placeOnline.place ?? ''),
+        appliedLinkForAll: [date?.placeOnline.appliedLinkForAll ?? false],
+        appliedPlaceForAll: [date?.placeOnline.appliedPlaceForAll ?? false]
+      })
+    });
+  }
+
+  transformDatesFormToDates(form: DateInformation[]): Dates[] {
+    return form
+      .map((value) => {
+        const { date, endTime, startTime } = value.day;
+        const { onlineLink, place, coordinates } = value.placeOnline;
+
+        const dateObject = new Date(date);
+
+        if (isNaN(dateObject.getTime())) {
+          return;
+        }
+
+        let [hours, minutes] = startTime.split(':');
+        dateObject.setHours(parseInt(hours, 10));
+        dateObject.setMinutes(parseInt(minutes, 10));
+        const startDate = dateObject.toISOString();
+
+        [hours, minutes] = endTime.split(':');
+        dateObject.setHours(parseInt(hours, 10));
+        dateObject.setMinutes(parseInt(minutes, 10));
+        const finishDate = dateObject.toISOString();
+
+        const dates: Dates = {
+          startDate,
+          finishDate,
+          id: undefined
+        };
+        if (onlineLink) {
+          dates.onlineLink = onlineLink;
+        }
+        if (place) {
+          dates.coordinates = {
+            latitude: coordinates.lat,
+            longitude: coordinates.lng
+          };
+        }
+        return dates;
+      })
+      .filter(Boolean);
+  }
+
+  prepareEventForSubmit(eventForm: EventForm, eventId: number, isUpdating: boolean) {
+    const { eventInformation, dateInformation } = eventForm;
+    const { open, tags, description, title, images } = eventInformation;
+    const dates: Dates[] = this.transformDatesFormToDates(dateInformation);
+    let sendEventDto: EventDTO = {
+      title,
+      description: description,
+      open,
+      tags,
+      datesLocations: dates
+    };
+
+    if (isUpdating) {
+      const currentImages = (eventForm?.eventInformation?.images || []).filter((value) => !value.file).map((value) => value.url);
+      sendEventDto = {
+        ...sendEventDto,
+        additionalImages: currentImages.slice(1),
+        id: eventId,
+        titleImage: currentImages[0]
+      };
+    }
+    const formData: FormData = new FormData();
+    const stringifyDataToSend = JSON.stringify(sendEventDto);
+    const dtoName = isUpdating ? 'eventDto' : 'addEventDtoRequest';
+
+    formData.append(dtoName, stringifyDataToSend);
+
+    images.forEach((item) => {
+      if (item.file) {
+        formData.append('images', item.file);
+      }
+    });
+
+    return formData;
   }
 
   getEventPreview(event: EventForm): PagePreviewDTO {
